homeworkweek12

+++ app/Http/Controllers/SupplierController.php
@@ -95,11 +95,7 @@ class SupplierController extends Controller
      */
     public function destroy(Supplier $supplier)
     {
-        // try{
-
-        // } catch (){
-
-        // }
+        $this->authorize('delete-permission', $supplier);
         $supplier->delete();
         return redirect()->route('suppliers.index')->with('status', 'Data berhasil dihapus');
     }
